refactor(storybook): share extractFunctionBody across verifyStorybook* guards (storybook/t-022)
verifyStorybookAnswerInputPreservationGuard, verifyStorybookRestartScenarioFrameGuard,
and verifyStorybookRestartInputCompletenessGuard each hand-rolled their own identical
brace-matching extractFunctionBody(content, name) helper, differing only in whether
the signature regex required `async function name(` vs plain `function name(` (and
one allowed leading indentation for a function nested inside a defineStore() setup
body).

Factored one shared implementation into utils/scripts/lib/extractTsFunctionBody.mjs
whose signature regex covers both shapes (optional `async`, optional leading
whitespace) and accepts an optional path/notFoundHint for guard-specific error text.
Updated all three guards to import it instead of keeping their own copy.

Verified: all 9 verifyStorybook*.mjs guards still pass. Each of the three touched
guards still fails correctly on its own documented regression case (git-stash
round-trip: dropping the restored answerInput assignment, dropping the forwarded
scenario field, and dropping it again for the completeness guard). eslint and
prettier --check clean on all touched files. vue-tsc --noEmit clean repo-wide.

Kaizen from storybook/t-021 (PR #1893).
